SlickOffsetStore uses Actor TestKit
